Finally fixed the traits import, thanks to a CCP fix and appropraite bacpac update from Steve Runoken (sp?)

+++ tools/XmlGenerator/Models/staStations.cs
@@ -12,7 +12,7 @@ namespace EVEMon.XmlGenerator.Models
         [DatabaseGenerated(DatabaseGeneratedOption.None)]
         public long stationID { get; set; }
 
-        public int? security { get; set; }
+        public double? security { get; set; }
 
         public double? dockingCostPerVolume { get; set; }
